fix: talk test
