+++ tests/IntegrationTests/Talk/TalkCommands.test.ts
@@ -1,782 +1,26 @@
-import { AppConfig } from "@/src/entities/config/AppConfig";
-import { Thread_Exclude_Prefix, Thread_Fetch_Nom } from "@/src/entities/constants/Thread";
-import type { ChatAIMessageDto } from "@/src/entities/dto/ChatAIMessageDto";
-import { ThreadDto } from "@/src/entities/dto/ThreadDto";
 import { CommunityCategoryType } from "@/src/entities/vo/CommunityCategoryType";
-import { CommunityId } from "@/src/entities/vo/CommunityId";
 import { PersonalityId } from "@/src/entities/vo/PersonalityId";
 import { ThreadCategoryType } from "@/src/entities/vo/ThreadCategoryType";
-import { ThreadMessageId } from "@/src/entities/vo/ThreadMessageId";
-import { ThreadMetadata } from "@/src/entities/vo/ThreadMetadata";
 import { TalkCommandHandler } from "@/src/handlers/discord.js/commands/TalkCommandHandler";
-import { AIReplyHandler } from "@/src/handlers/discord.js/events/AIReplyHandler";
-import type { IChatAILogic } from "@/src/logics/Interfaces/logics/IChatAILogic";
-import type { ICommunityLogic } from "@/src/logics/Interfaces/logics/ICommunityLogic";
-import { ThreadLogic } from "@/src/logics/ThreadLogic";
-import { DiscordTextPresenter } from "@/src/presenter/DiscordTextPresenter";
 import { CommunityRepositoryImpl } from "@/src/repositories/sequelize-mysql/CommunityRepositoryImpl";
 import { ContextRepositoryImpl } from "@/src/repositories/sequelize-mysql/ContextRepositoryImpl";
 import { MysqlConnector } from "@/src/repositories/sequelize-mysql/MysqlConnector";
 import { PersonalityContextRepositoryImpl } from "@/src/repositories/sequelize-mysql/PersonalityContextRepositoryImpl";
 import { PersonalityRepositoryImpl } from "@/src/repositories/sequelize-mysql/PersonalityRepositoryImpl";
 import { ThreadRepositoryImpl } from "@/src/repositories/sequelize-mysql/ThreadRepositoryImpl";
-import { mockMessage } from "@/tests/fixtures/discord.js/MockMessage";
 import { mockSlashCommand, waitUntilReply } from "@/tests/fixtures/discord.js/MockSlashCommand";
-import { TestDiscordServer } from "@/tests/fixtures/discord.js/TestDiscordServer";
 import { expect } from "chai";
-import type { TextChannel } from "discord.js";
 import { anything, instance, mock, verify, when } from "ts-mockito";
 
-// ============================================================
-// 型定義
-// ============================================================
-
-/**
- * スレッドメタデータの型定義
- */
-interface ThreadMetadataContent {
-	persona_role: string;
-	speaking_style_rules: string;
-	response_directives: string;
-	emotion_model: string;
-	notes: string;
-	input_scope: string;
-}
-
-/**
- * スレッドデータの型定義（Repositoryから取得したデータ）
- */
-interface ThreadRecord {
-	communityId: number | { toString(): string };
-	messageId: string | { toString(): string };
-	categoryType: number;
-	metadata: ThreadMetadataContent | JSON | null;
-	toDto?: () => ThreadDto;
-}
-
-// ============================================================
-// 共通テスト定数
-// ============================================================
-
-// テスト用のguildId（MockSlashCommandで使用される値と一致させる）
-const TEST_GUILD_ID = "12345";
-const TEST_THREAD_ID = "67890";
-const TEST_USER_ID = "98765";
-const TEST_BOT_ID = AppConfig.discord.clientId;
-
-// ============================================================
-// 共通モック生成ヘルパー関数
-// ============================================================
-
-/**
- * テスト用のスレッドメタデータを生成する
- * @param overrides - 上書きするプロパティ
- * @returns スレッドメタデータオブジェクト
- */
-function createTestMetadata(overrides: Partial<ThreadMetadataContent> = {}): ThreadMetadataContent {
-	return {
-		persona_role: "テスト役割",
-		speaking_style_rules: "テストスタイル",
-		response_directives: "テスト指示",
-		emotion_model: "テスト感情",
-		notes: "テスト注釈",
-		input_scope: "テスト範囲",
-		...overrides,
-	};
-}
-
-/**
- * テスト用のスレッドデータをリポジトリに作成する
- * @param options - オプションパラメータ
- * @returns 作成されたスレッドデータ
- */
-async function createTestThread(
-	options: {
-		communityId?: number;
-		messageId?: string;
-		categoryType?: number;
-		metadata?: object;
-	} = {},
-) {
-	return await ThreadRepositoryImpl.create({
-		communityId: options.communityId ?? 1,
-		messageId: options.messageId ?? TEST_THREAD_ID,
-		categoryType: options.categoryType ?? ThreadCategoryType.CATEGORY_TYPE_CHATGPT.getValue(),
-		metadata: options.metadata ?? createTestMetadata(),
-	});
-}
-
-/**
- * テスト用のThreadDtoを生成する
- * @param options - オプションパラメータ
- * @returns ThreadDtoインスタンス
- */
-function createTestThreadDto(
-	options: {
-		communityId?: number;
-		messageId?: string;
-		categoryType?: ThreadCategoryType;
-		metadata?: object;
-	} = {},
-): ThreadDto {
-	return new ThreadDto(
-		new CommunityId(options.communityId ?? 1),
-		new ThreadMessageId(options.messageId ?? TEST_THREAD_ID),
-		options.categoryType ?? ThreadCategoryType.CATEGORY_TYPE_CHATGPT,
-		new ThreadMetadata((options.metadata ?? createTestMetadata()) as unknown as JSON),
-	);
-}
-
-/**
- * テスト用のメッセージコレクションモックを生成する
- * @param messages - メッセージ配列
- * @returns メッセージコレクションモック
- */
-function createMessageCollectionMock(
-	messages: Array<{
-		id?: string;
-		author: { bot: boolean; id?: string };
-		content: string;
-	}>,
-) {
-	return {
-		reverse: () => messages,
-		map: function (callback: (msg: (typeof messages)[0]) => unknown) {
-			return this.reverse().map(callback);
-		},
-	};
-}
-
-/**
- * テスト用のチャンネルモックを生成・設定する
- * @param options - オプションパラメータ
- * @returns 設定済みのチャンネルモック
- */
-function createChannelMock(
-	options: {
-		isThread?: boolean;
-		guildId?: string;
-		threadId?: string;
-		ownerId?: string;
-		messageCollection?: ReturnType<typeof createMessageCollectionMock>;
-	} = {},
-) {
-	const channelMock = mock<any>();
-	when(channelMock.isThread()).thenReturn(options.isThread ?? true);
-	when(channelMock.guildId).thenReturn(options.guildId ?? TEST_GUILD_ID);
-	when(channelMock.id).thenReturn(options.threadId ?? TEST_THREAD_ID);
-	when(channelMock.ownerId).thenReturn(options.ownerId ?? TEST_BOT_ID);
-	when(channelMock.sendTyping()).thenResolve();
-	when(channelMock.messages).thenReturn({
-		fetch: (fetchOptions?: { limit: number }) => {
-			if (fetchOptions) {
-				expect(fetchOptions).to.deep.equal({ limit: Thread_Fetch_Nom });
-			}
-			return Promise.resolve(options.messageCollection ?? createMessageCollectionMock([]));
-		},
-	});
-	return channelMock;
-}
-
-/**
- * TextChannel用のモックを生成する
- * @returns 設定済みのTextChannelモック
- */
-function createTextChannelMock() {
-	const channelMock = mock<TextChannel>();
-	when(channelMock.threads).thenReturn({
-		create: async () => ({}),
-	} as any);
-	return channelMock;
-}
-
-// ============================================================
-// イベント登録テスト用ヘルパー関数
-// ============================================================
-
-/**
- * Discord interactionイベントを発行する
- * @param commandMock - コマンドモック
- */
-async function emitInteractionEvent(commandMock: any): Promise<void> {
-	const TEST_CLIENT = await TestDiscordServer.getClient();
-	TEST_CLIENT.emit("interactionCreate", instance(commandMock));
-}
-
-/**
- * AIReplyHandlerでメッセージを処理する
- * @param handler - AIReplyHandlerインスタンス
- * @param messageMock - メッセージモック
- */
-async function handleAIReplyEvent(handler: AIReplyHandler, messageMock: any): Promise<void> {
-	await handler.handle(instance(messageMock));
-}
-
-/**
- * コマンド実行結果の期待値オプション
- */
-interface CommandTestExpectation {
-	/** 応答メッセージの検証 */
-	expectReply?: {
-		/** 応答が呼ばれる回数 (デフォルト: 1) */
-		times?: number;
-		/** 応答が呼ばれないことを期待 */
-		never?: boolean;
-		/** 応答メッセージの内容 */
-		content?: string;
-	};
-	/** スレッド作成の検証 */
-	expectThread?: {
-		/** スレッドが作成されることを期待 */
-		created?: boolean;
-		/** 期待するスレッド数 */
-		count?: number;
-		/** スレッドデータの検証 */
-		data?: {
-			communityId?: string;
-			messageId?: string;
-			categoryType?: number;
-			metadata?: object;
-		};
-	};
-	/** 待機時間（ミリ秒） */
-	waitTime?: number;
-}
-
-/**
- * TalkCommandのテスト実行ヘルパー関数
- * コマンド発行からレスポンス検証までを一括で行う
- * @param commandOptions - コマンドオプション
- * @param mockSetup - モック設定
- * @param expectations - 期待値
- * @returns テスト結果
- */
-async function executeTalkCommandTest(
-	commandOptions: {
-		title: string | null;
-		type: number;
-	},
-	mockSetup: {
-		guildId?: string;
-		setupChannel?: boolean;
-		replyResponse?: any;
-	} = {},
-	expectations: CommandTestExpectation = {},
-): Promise<{
-	commandMock: any;
-	channelMock: any;
-}> {
-	const guildId = mockSetup.guildId ?? TEST_GUILD_ID;
-
-	// コマンドモックの設定
-	const commandMock = mockSlashCommand("talk", commandOptions, { guildId });
-
-	// チャンネルモックの設定
-	let channelMock: any = null;
-	if (mockSetup.setupChannel !== false) {
-		channelMock = createTextChannelMock();
-		when(commandMock.channel).thenReturn(instance(channelMock));
-	}
-
-	// レスポンスモックの設定
-	if (mockSetup.replyResponse !== undefined) {
-		when(commandMock.reply(anything())).thenResolve(mockSetup.replyResponse);
-	}
-
-	// イベント発行
-	await emitInteractionEvent(commandMock);
-
-	// 待機
-	if (expectations.waitTime) {
-		await new Promise((resolve) => setTimeout(resolve, expectations.waitTime));
-	}
-
-	// 応答の検証
-	if (expectations.expectReply) {
-		if (expectations.expectReply.never) {
-			verify(commandMock.reply(anything())).never();
-		} else if (expectations.expectReply.content) {
-			verify(commandMock.reply(expectations.expectReply.content)).times(expectations.expectReply.times ?? 1);
-		} else {
-			verify(commandMock.reply(anything())).times(expectations.expectReply.times ?? 1);
-		}
-	}
-
-	// スレッド作成の検証
-	if (expectations.expectThread) {
-		if (expectations.expectThread.created === false || expectations.expectThread.count === 0) {
-			await assertNoThreadsCreated();
-		} else if (expectations.expectThread.count !== undefined) {
-			await assertThreadCount(expectations.expectThread.count);
-		}
-	}
-
-	return { commandMock, channelMock };
-}
-
-/**
- * AIReplyHandlerテスト結果の型定義
- */
-interface AIReplyTestResult {
-	handler: AIReplyHandler;
-	messageMock: ReturnType<typeof mockMessage>;
-	channelMock: ReturnType<typeof createChannelMock>;
-	communityLogicMock: ICommunityLogic;
-	threadLogicMock: ThreadLogic;
-	chatAILogicMock: IChatAILogic;
-}
-
-/**
- * AIReplyHandlerの動作検証オプション
- */
-interface AIReplyTestExpectation {
-	/** AIロジックが呼ばれる回数 */
-	expectAICall?: {
-		times?: number;
-		never?: boolean;
-	};
-	/** 応答の検証 */
-	expectReply?: {
-		times?: number;
-		never?: boolean;
-		content?: string;
-	};
-	/** sendTypingの検証 */
-	expectTyping?: {
-		times?: number;
-		never?: boolean;
-	};
-}
-
-/**
- * AIReplyHandlerのテストを実行し、結果を検証する
- * @param options - テストオプション
- * @param expectations - 期待値
- * @returns テスト結果
- */
-async function executeAIReplyTestWithVerification(
-	options: {
-		userId?: string;
-		content?: string;
-		messageHistory?: Array<{
-			id?: string;
-			author: { bot: boolean; id?: string };
-			content: string;
-		}>;
-		threadDto?: ThreadDto;
-		replyResponse?: string;
-		replyCallback?: (prompt: unknown, context: ChatAIMessageDto[]) => Promise<string>;
-		isThread?: boolean;
-		ownerId?: string;
-	} = {},
-	expectations: AIReplyTestExpectation = {},
-): Promise<AIReplyTestResult> {
-	// テストを実行
-	const result = await executeAIReplyTest(options);
-
-	// AIロジック呼び出しの検証
-	if (expectations.expectAICall) {
-		if (expectations.expectAICall.never) {
-			verify(result.chatAILogicMock.replyTalk(anything(), anything())).never();
-		} else {
-			verify(result.chatAILogicMock.replyTalk(anything(), anything())).times(expectations.expectAICall.times ?? 1);
-		}
-	}
-
-	// 応答の検証
-	if (expectations.expectReply) {
-		if (expectations.expectReply.never) {
-			verify(result.messageMock.reply(anything())).never();
-		} else if (expectations.expectReply.content) {
-			verify(result.messageMock.reply(expectations.expectReply.content)).times(expectations.expectReply.times ?? 1);
-		} else {
-			verify(result.messageMock.reply(anything())).times(expectations.expectReply.times ?? 1);
-		}
-	}
-
-	// sendTypingの検証
-	if (expectations.expectTyping) {
-		if (expectations.expectTyping.never) {
-			verify(result.channelMock.sendTyping()).never();
-		} else {
-			verify(result.channelMock.sendTyping()).times(expectations.expectTyping.times ?? 1);
-		}
-	}
-
-	return result;
-}
-
-/**
- * メッセージフィルタリングテスト用のセットアップヘルパー
- * @param options - オプション
- * @returns 設定済みのモックとハンドラー
- */
-async function setupMessageFilteringTest(
-	options: {
-		isBot?: boolean;
-		isThread?: boolean;
-		ownerId?: string;
-		threadId?: string;
-		categoryType?: ThreadCategoryType;
-	} = {},
-): Promise<{
-	handler: AIReplyHandler;
-	messageMock: any;
-	channelMock: any;
-}> {
-	const { handler } = createAIReplyHandlerWithMocks({
-		threadDto: options.categoryType
-			? createTestThreadDto({
-					messageId: options.threadId ?? TEST_THREAD_ID,
-					categoryType: options.categoryType,
-				})
-			: undefined,
-	});
-
-	const messageMock = mockMessage(TEST_USER_ID, false, options.isBot ?? false);
-	const channelMock = createChannelMock({
-		isThread: options.isThread ?? true,
-		threadId: options.threadId ?? TEST_THREAD_ID,
-		ownerId: options.ownerId ?? TEST_BOT_ID,
-	});
-
-	when(messageMock.channel).thenReturn(instance(channelMock));
-	when(messageMock.reply(anything())).thenResolve();
-
-	return { handler, messageMock, channelMock };
-}
-
-/**
- * AIReplyHandler初期化オプションの型定義
- */
-interface AIReplyHandlerMockOptions {
-	/** ThreadDtoを返すモック設定 */
-	threadDto?: ThreadDto;
-	/** AI応答文字列 */
-	replyResponse?: string;
-	/** AI応答コールバック（詳細な検証用） */
-	replyCallback?: (prompt: unknown, context: ChatAIMessageDto[]) => Promise<string>;
-	/** ThreadLogic.findの振る舞いをカスタマイズ */
-	findCallback?: (communityId: CommunityId, messageId: ThreadMessageId) => Promise<ThreadDto | undefined>;
-	/** AI応答でエラーをスロー */
-	replyThrowError?: Error;
-}
-
-/**
- * AIReplyHandlerのインスタンスを作成し、モックを設定する
- * @param options - オプションパラメータ
- * @returns 設定済みのAIReplyHandlerとモック
- */
-function createAIReplyHandlerWithMocks(options: AIReplyHandlerMockOptions = {}) {
-	const handler = new AIReplyHandler();
-
-	// CommunityLogicのモックを作成
-	const communityLogicMock = mock<ICommunityLogic>();
-	// @ts-ignore - privateフィールドにアクセスするため
-	handler.CommunityLogic = instance(communityLogicMock);
-	when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-	// ThreadLogicのモックを作成
-	const threadLogicMock = mock<ThreadLogic>();
-	// @ts-ignore - privateフィールドにアクセスするため
-	handler.threadLogic = instance(threadLogicMock);
-
-	if (options.findCallback) {
-		when(threadLogicMock.find(anything(), anything())).thenCall(options.findCallback);
-	} else if (options.threadDto) {
-		when(threadLogicMock.find(anything(), anything())).thenResolve(options.threadDto);
-	}
-
-	// ChatAILogicのモックを作成
-	const chatAILogicMock = mock<IChatAILogic>();
-	// @ts-ignore - privateフィールドにアクセスするため
-	handler.chatAILogic = instance(chatAILogicMock);
-
-	if (options.replyThrowError) {
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenThrow(options.replyThrowError);
-	} else if (options.replyCallback) {
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall(options.replyCallback);
-	} else if (options.replyResponse !== undefined) {
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve(options.replyResponse);
-	}
-
-	return {
-		handler,
-		communityLogicMock,
-		threadLogicMock,
-		chatAILogicMock,
-	};
-}
-
-/**
- * TalkCommandHandlerのインスタンスを作成する
- * @returns 設定済みのTalkCommandHandler
- */
-function createTalkCommandHandler() {
-	return new TalkCommandHandler();
-}
-
-// ============================================================
-// Repositoryテスト用ヘルパー関数
-// ============================================================
-
-/**
- * 全スレッドを取得する
- * @returns スレッドの配列
- */
-async function findAllThreads() {
-	return await ThreadRepositoryImpl.findAll();
-}
-
-/**
- * メッセージIDでスレッドを検索する
- * @param messageId - メッセージID
- * @param communityId - コミュニティID（デフォルト: 1）
- * @returns スレッドまたはnull
- */
-async function findThreadByMessageId(messageId: string, communityId = 1) {
-	return await ThreadRepositoryImpl.findOne({
-		where: {
-			communityId,
-			messageId,
-		},
-	});
-}
-
-/**
- * スレッドが作成されていないことを検証する
- */
-async function assertNoThreadsCreated() {
-	const threads = await findAllThreads();
-	expect(threads.length).to.eq(0);
-}
-
-/**
- * スレッド数を検証する
- * @param expectedCount - 期待するスレッド数
- */
-async function assertThreadCount(expectedCount: number) {
-	const threads = await findAllThreads();
-	expect(threads.length).to.eq(expectedCount);
-}
-
-/**
- * スレッドの基本データを検証する
- * @param thread - 検証対象のスレッド
- * @param expected - 期待値
- */
-function verifyThreadData(
-	thread: ThreadRecord,
-	expected: {
-		communityId?: string;
-		messageId?: string;
-		categoryType?: number;
-	},
-): void {
-	if (expected.communityId !== undefined) {
-		expect(thread.communityId.toString()).to.eq(expected.communityId);
-	}
-	if (expected.messageId !== undefined) {
-		expect(thread.messageId.toString()).to.eq(expected.messageId);
-	}
-	if (expected.categoryType !== undefined) {
-		expect(thread.categoryType).to.eq(expected.categoryType);
-	}
-}
-
-/**
- * スレッドのメタデータを検証する
- * @param thread - 検証対象のスレッド
- * @param expectedMetadata - 期待するメタデータ（部分一致）
- */
-function verifyThreadMetadata(
-	thread: ThreadRecord,
-	expectedMetadata?: Partial<ThreadMetadataContent>,
-): void {
-	const metadata = thread.metadata;
-	expect(metadata).to.not.be.null;
-	expect(metadata).to.be.an("object");
-
-	const metadataObj = JSON.parse(JSON.stringify(metadata)) as ThreadMetadataContent;
-
-	// 基本的なプロパティの存在確認
-	expect(metadataObj).to.have.property("persona_role");
-	expect(metadataObj).to.have.property("speaking_style_rules");
-	expect(metadataObj).to.have.property("response_directives");
-	expect(metadataObj).to.have.property("emotion_model");
-	expect(metadataObj).to.have.property("notes");
-	expect(metadataObj).to.have.property("input_scope");
-
-	// 期待値が指定されている場合は値も検証
-	if (expectedMetadata) {
-		for (const [key, value] of Object.entries(expectedMetadata)) {
-			if (value !== undefined) {
-				expect(metadataObj[key as keyof ThreadMetadataContent]).to.eq(value);
-			}
-		}
-	}
-}
-
-/**
- * スレッドが存在し、データが正しいことを検証する
- * @param messageId - メッセージID
- * @param expected - 期待値
- * @param communityId - コミュニティID（デフォルト: 1）
- */
-async function assertThreadExistsWithData(
-	messageId: string,
-	expected: {
-		communityId?: string;
-		categoryType?: number;
-		metadata?: Partial<{
-			persona_role: string;
-			speaking_style_rules: string;
-			response_directives: string;
-			emotion_model: string;
-			notes: string;
-			input_scope: string;
-		}>;
-	},
-	communityId = 1,
-) {
-	const thread = await findThreadByMessageId(messageId, communityId);
-	expect(thread).to.not.be.null;
-
-	if (thread) {
-		verifyThreadData(thread, {
-			communityId: expected.communityId,
-			messageId,
-			categoryType: expected.categoryType,
-		});
-
-		if (expected.metadata !== undefined) {
-			verifyThreadMetadata(thread, expected.metadata);
-		}
-	}
-}
-
-/**
- * メッセージモックとチャンネルモックを一括設定する
- * @param options - オプションパラメータ
- * @returns 設定済みのメッセージモックとチャンネルモック
- */
-function setupMessageWithChannel(
-	options: {
-		userId?: string;
-		content?: string;
-		isThread?: boolean;
-		guildId?: string;
-		threadId?: string;
-		ownerId?: string;
-		messageCollection?: ReturnType<typeof createMessageCollectionMock>;
-	} = {},
-) {
-	const messageMock = mockMessage(options.userId ?? TEST_USER_ID);
-	if (options.content !== undefined) {
-		when(messageMock.content).thenReturn(options.content);
-	}
-
-	const channelMock = createChannelMock({
-		isThread: options.isThread,
-		guildId: options.guildId,
-		threadId: options.threadId,
-		ownerId: options.ownerId,
-		messageCollection: options.messageCollection,
-	});
-
-	when(messageMock.channel).thenReturn(instance(channelMock));
-	when(messageMock.reply(anything())).thenResolve();
-
-	return { messageMock, channelMock };
-}
-
-/**
- * executeAIReplyTestのオプション型定義
- */
-interface ExecuteAIReplyTestOptions {
-	/** ユーザーID */
-	userId?: string;
-	/** メッセージ内容 */
-	content?: string;
-	/** メッセージ履歴 */
-	messageHistory?: Array<{
-		id?: string;
-		author: { bot: boolean; id?: string };
-		content: string;
-	}>;
-	/** スレッドDTO */
-	threadDto?: ThreadDto;
-	/** AI応答文字列 */
-	replyResponse?: string;
-	/** AI応答コールバック */
-	replyCallback?: (prompt: unknown, context: ChatAIMessageDto[]) => Promise<string>;
-}
-
-/**
- * AIReplyHandlerのテストを簡略化するためのヘルパー関数
- * Handler作成、メッセージ設定、ハンドル実行を一括で行う
- * @param options - オプションパラメータ
- * @returns テスト結果の検証用オブジェクト
- */
-async function executeAIReplyTest(options: ExecuteAIReplyTestOptions = {}): Promise<AIReplyTestResult> {
-	// AIReplyHandlerとモックを作成
-	const { handler, communityLogicMock, threadLogicMock, chatAILogicMock } = createAIReplyHandlerWithMocks({
-		threadDto: options.threadDto,
-		replyResponse: options.replyResponse,
-		replyCallback: options.replyCallback,
-	});
-
-	// メッセージとチャンネルを設定
-	const messageCollection = options.messageHistory ? createMessageCollectionMock(options.messageHistory) : undefined;
-
-	const { messageMock, channelMock } = setupMessageWithChannel({
-		userId: options.userId,
-		content: options.content,
-		messageCollection,
-	});
-
-	// ハンドル実行
-	await handleAIReplyEvent(handler, messageMock);
-
-	return {
-		handler,
-		messageMock,
-		channelMock,
-		communityLogicMock,
-		threadLogicMock,
-		chatAILogicMock,
-	};
-}
-
-/**
- * Commandテストを簡略化するためのヘルパー関数
- * コマンドモック作成、イベント発行、応答待機を一括で行う
- * @param commandName - コマンド名
- * @param commandOptions - コマンドオプション
- * @param mockOptions - モックオプション
- * @returns テスト結果の検証用オブジェクト
- */
-async function executeCommandTest(
-	commandName: string,
-	commandOptions: Record<string, unknown>,
-	mockOptions: {
-		guildId?: string;
-		setupChannel?: boolean;
-	} = {},
-) {
-	const commandMock = mockSlashCommand(commandName, commandOptions, {
-		guildId: mockOptions.guildId ?? TEST_GUILD_ID,
-	});
-
-	if (mockOptions.setupChannel !== false) {
-		const channelMock = createTextChannelMock();
-		when(commandMock.channel).thenReturn(instance(channelMock));
-	}
-
-	await emitInteractionEvent(commandMock);
-
-	return { commandMock };
-}
+import {
+	TEST_GUILD_ID,
+	createTextChannelMock,
+	emitInteractionEvent,
+	executeTalkCommandTest,
+	assertNoThreadsCreated,
+	assertThreadCount,
+	assertThreadExistsWithData,
+} from "./TalkHelper.test";
 
 describe("Test Talk Commands", function (this: Mocha.Suite) {
 	// テストのタイムアウト時間を延長（60秒）
@@ -1116,1910 +360,5 @@ describe("Test Talk Commands", function (this: Mocha.Suite) {
 			{},
 			{ expectThread: { count: 0 } },
 		);
-	});
-
-	/**
-	 * AIReplyHandlerのテスト
-	 */
-
-	/**
-	 * メッセージフィルタリングテスト
-	 * - Bot自身の発言を無視できているかを確認
-	 * - スレッド以外のチャンネルからのメッセージが無視されるか
-	 * - 他ユーザーが所有するスレッドが除外対象になるか
-	 * - カスタムカテゴリ（CHATGPT以外）のスレッドで無視されるか
-	 */
-	it("test AIReplyHandler message filtering", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testOtherThreadId = "67891";
-		const testNonChatGPTThreadId = "67892";
-
-		// テスト用のスレッドデータを作成
-		// 1. ChatGPTカテゴリのスレッド（Botが所有）
-		await createTestThread();
-
-		// 2. ChatGPTカテゴリのスレッド（他ユーザーが所有）
-		await createTestThread({ messageId: testOtherThreadId });
-
-		// 3. 非ChatGPTカテゴリのスレッド（Botが所有）
-		await createTestThread({
-			messageId: testNonChatGPTThreadId,
-			categoryType: ThreadCategoryType.CATEGORY_TYPE_DEEPL.getValue(),
-		});
-
-		// ヘルパー関数を使用してAIReplyHandlerとモックを作成
-		const { handler, threadLogicMock } = createAIReplyHandlerWithMocks();
-
-		// テストケース1: Bot自身の発言を無視できているか
-		const botMessageMock = mockMessage(TEST_BOT_ID, false, true);
-		when(botMessageMock.channel).thenReturn({
-			isThread: () => true,
-			communityId: 1,
-			id: TEST_THREAD_ID,
-			ownerId: TEST_BOT_ID,
-			sendTyping: () => Promise.resolve(),
-			messages: {
-				fetch: () => Promise.resolve([]),
-			},
-		} as any);
-
-		await handleAIReplyEvent(handler, botMessageMock);
-		verify(botMessageMock.reply(anything())).never();
-
-		// テストケース2: スレッド以外のチャンネルからのメッセージが無視されるか
-		const { messageMock: nonThreadMessageMock } = setupMessageWithChannel({
-			isThread: false,
-			threadId: "12345",
-		});
-
-		await handleAIReplyEvent(handler, nonThreadMessageMock);
-		verify(nonThreadMessageMock.reply(anything())).never();
-
-		// テストケース3: 他ユーザーが所有するスレッドが除外対象になるか
-		const { messageMock: otherOwnerMessageMock } = setupMessageWithChannel({
-			threadId: testOtherThreadId,
-			ownerId: TEST_USER_ID, // Botではなく他のユーザーがオーナー
-		});
-
-		await handleAIReplyEvent(handler, otherOwnerMessageMock);
-		verify(otherOwnerMessageMock.reply(anything())).never();
-
-		// テストケース4: カスタムカテゴリ（CHATGPT以外）のスレッドで無視されるか
-		when(threadLogicMock.find(anything(), anything())).thenResolve(
-			createTestThreadDto({
-				messageId: testNonChatGPTThreadId,
-				categoryType: ThreadCategoryType.CATEGORY_TYPE_DEEPL, // ChatGPT以外のカテゴリ
-			}),
-		);
-
-		const { messageMock: nonChatGPTMessageMock } = setupMessageWithChannel({
-			threadId: testNonChatGPTThreadId,
-		});
-
-		await handleAIReplyEvent(handler, nonChatGPTMessageMock);
-		verify(nonChatGPTMessageMock.reply(anything())).never();
-	});
-
-	/**
-	 * [ThreadExcludePrefix] 除外プレフィックスの検証
-	 * - ';' を付けたメッセージには反応しないか
-	 */
-	it("test semicolon-prefixed messages are ignored", async function (this: Mocha.Context) {
-		this.timeout(10_000);
-
-		// テスト用スレッドを作成
-		await createTestThread();
-
-		// ヘルパー関数を使用してテストを実行
-		const { chatAILogicMock, messageMock } = await executeAIReplyTest({
-			content: `${Thread_Exclude_Prefix}無視してほしいメッセージ`,
-			threadDto: createTestThreadDto(),
-		});
-
-		verify(chatAILogicMock.replyTalk(anything(), anything())).never();
-		verify(messageMock.reply(anything())).never();
-	});
-
-	/**
-	 * [ThreadExcludePrefix] 除外プレフィックスの検証
-	 * - 履歴取得時に ';' 付きメッセージがコンテキストに含まれないか
-	 */
-	it("test semicolon-prefixed messages are excluded from context history", async function (this: Mocha.Context) {
-		this.timeout(10_000);
-
-		// テスト用スレッドを作成
-		await createTestThread();
-
-		// メッセージ履歴（除外プレフィックス付きメッセージを含む）
-		const messageHistory = [
-			{ id: "msg1", author: { bot: false, id: TEST_USER_ID }, content: "こんにちは" },
-			{ id: "msg2", author: { bot: false, id: TEST_USER_ID }, content: `${Thread_Exclude_Prefix}除外メッセージ` },
-			{ id: "msg3", author: { bot: true, id: TEST_BOT_ID }, content: "前回の応答" },
-			{ id: "msg4", author: { bot: false, id: TEST_USER_ID }, content: "質問です" },
-		];
-
-		// ヘルパー関数を使用してテストを実行
-		const { chatAILogicMock, messageMock } = await executeAIReplyTest({
-			content: "質問です",
-			messageHistory,
-			threadDto: createTestThreadDto(),
-			replyCallback: (prompt, context) => {
-				const contents = context.map((entry: ChatAIMessageDto) => entry.content.getValue());
-				expect(contents).to.deep.equal(["こんにちは", "前回の応答", "質問です"]);
-				return Promise.resolve("テスト応答");
-			},
-		});
-
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-		verify(messageMock.reply(anything())).once();
-	});
-
-	/**
-	 * [ThreadSearch] スレッド検索機能の検証
-	 * - ThreadLogic.find が適切な引数で呼ばれるか
-	 * - CommunityId および ThreadMessageId が正しい形式で生成されるか
-	 * - 対象スレッドが存在しないケースでのハンドリングが正しいか
-	 */
-	it("test ThreadLogic.find functionality", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = 12345;
-		const testThreadId = 67890;
-		const testNonExistThreadId = 99999;
-		const testUserId = 98765;
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId.toString() });
-
-		// ThreadLogicのインスタンスを作成
-		const threadLogic = new ThreadLogic();
-		// @ts-ignore - privateフィールドにアクセスするため
-		const threadRepositoryMock = mock<ThreadRepositoryImpl>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		threadLogic.threadRepository = instance(threadRepositoryMock);
-		// @ts-ignore - privateフィールドにアクセスするため
-		threadLogic.transaction = {
-			startTransaction: async (callback: () => Promise<any>) => {
-				return await callback();
-			},
-		};
-
-		// 正常系: 存在するスレッドを検索
-		when(threadRepositoryMock.findByMessageId(anything(), anything())).thenCall(async (communityId: CommunityId, messageId: ThreadMessageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(Number(messageId.getValue())).to.equal(testThreadId);
-
-			// 実際のデータベースからスレッドを取得
-			return await ThreadRepositoryImpl.findOne({
-				where: {
-					communityId: 1,
-					messageId: messageId.getValue(),
-				},
-			}).then((res) => (res ? res.toDto() : undefined));
-		});
-
-		// スレッド検索の実行
-		const foundThread = await threadLogic.find(new CommunityId(1), new ThreadMessageId(testThreadId.toString()));
-
-		// 検索結果の検証
-		expect(foundThread).to.not.be.undefined;
-		if (foundThread) {
-			expect(foundThread.communityId.getValue()).to.equal(1);
-			expect(Number(foundThread.messageId.getValue())).to.equal(testThreadId);
-			expect(foundThread.categoryType.getValue()).to.equal(ThreadCategoryType.CATEGORY_TYPE_CHATGPT.getValue());
-
-			// メタデータの検証
-			const metadata = foundThread.metadata.getValue();
-			expect(metadata).to.have.property("persona_role", "テスト役割");
-			expect(metadata).to.have.property("speaking_style_rules", "テストスタイル");
-			expect(metadata).to.have.property("response_directives", "テスト指示");
-			expect(metadata).to.have.property("emotion_model", "テスト感情");
-			expect(metadata).to.have.property("notes", "テスト注釈");
-			expect(metadata).to.have.property("input_scope", "テスト範囲");
-		}
-
-		// 異常系: 存在しないスレッドを検索
-		when(threadRepositoryMock.findByMessageId(anything(), anything())).thenCall(async (communityId: CommunityId, messageId: ThreadMessageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(Number(messageId.getValue())).to.equal(testNonExistThreadId);
-
-			// 存在しないスレッドの場合はundefinedを返す
-			return undefined;
-		});
-
-		// 存在しないスレッドの検索実行
-		const notFoundThread = await threadLogic.find(new CommunityId(1), new ThreadMessageId(testNonExistThreadId.toString()));
-
-		// 検索結果の検証（存在しない場合はundefinedが返されるはず）
-		expect(notFoundThread).to.be.undefined;
-
-		// AIReplyHandlerでのスレッド検索の挙動を検証
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-		// @ts-ignore - privateフィールドにアクセスするため
-		const threadLogicForHandlerMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicForHandlerMock);
-
-		// 存在しないスレッドの場合
-		when(threadLogicForHandlerMock.find(anything(), anything())).thenResolve(undefined);
-
-		const messageMock = mockMessage(testUserId.toString());
-		when(messageMock.channel).thenReturn({
-			isThread: () => true,
-			communityId: 1,
-			id: testNonExistThreadId,
-			ownerId: AppConfig.discord.clientId,
-			sendTyping: () => Promise.resolve(),
-		} as any);
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(messageMock));
-
-		// スレッドが存在しない場合は処理がスキップされるため、replyは呼ばれないはず
-		verify(messageMock.reply(anything())).never();
-
-		// ThreadGuildIdとThreadMessageIdの生成と検証
-		const guildId = new CommunityId(testGuildId);
-		const messageId = new ThreadMessageId(testThreadId.toString());
-
-		expect(Number(guildId.getValue())).to.equal(testGuildId);
-		expect(Number(messageId.getValue())).to.equal(testThreadId);
-		expect(Number(guildId.getValue())).to.equal(testGuildId);
-		expect(Number(messageId.getValue())).to.equal(testThreadId);
-	});
-
-	/**
-	 * [TypingIndicator] タイピング表示の検証
-	 * - sendTyping が正しく呼ばれているか
-	 * - 入力検出直後などタイミング的に妥当な場所で動くか
-	 */
-	it("test typing indicator is shown at appropriate timing", async function (this: Mocha.Context) {
-		this.timeout(10_000);
-
-		// テスト用のスレッドデータを作成
-		await createTestThread();
-
-		// ヘルパー関数を使用してAIReplyHandlerとモックを作成
-		const { handler, threadLogicMock, chatAILogicMock } = createAIReplyHandlerWithMocks({
-			threadDto: createTestThreadDto(),
-			replyResponse: "テスト応答",
-		});
-
-		// メッセージ履歴のモック
-		const messageHistory = [{ id: "msg1", author: { bot: false, id: TEST_USER_ID }, content: "こんにちは" }];
-
-		// メッセージとチャンネルを設定
-		const { messageMock, channelMock } = setupMessageWithChannel({
-			messageCollection: createMessageCollectionMock(messageHistory),
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await handleAIReplyEvent(handler, messageMock);
-
-		// sendTypingが呼ばれることを確認
-		verify(channelMock.sendTyping()).once();
-
-		// 処理の順序を検証
-		verify(threadLogicMock.find(anything(), anything())).once();
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-		verify(messageMock.reply(anything())).once();
-	});
-
-	/**
-	 * [MessageHistory] メッセージ履歴取得と変換の検証
-	 * - channel.messages.fetch が `limit: 21` で正しく呼ばれるか
-	 * - メッセージ取得結果が時間順に逆順ソートされるか
-	 * - ユーザーメッセージが USER ロールへ、Botメッセージが ASSISTANT ロールへ正しく変換されるか
-	 * - 内容が ChatAIContent 構造として正しく渡せるか
-	 */
-	it("test message history retrieval and conversion", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMockForTest = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMockForTest);
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMockForTest.find(anything(), anything())).thenResolve(
-			new ThreadDto(
-				new CommunityId(1),
-				new ThreadMessageId(testThreadId),
-				ThreadCategoryType.CATEGORY_TYPE_CHATGPT,
-				new ThreadMetadata({
-					persona_role: "テスト役割",
-					speaking_style_rules: "テストスタイル",
-					response_directives: "テスト指示",
-					emotion_model: "テスト感情",
-					notes: "テスト注釈",
-					input_scope: "テスト範囲",
-				} as unknown as JSON),
-			),
-		);
-
-		// ChatAILogic.replyTalkメソッドのモック
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve("テスト応答");
-
-		// テスト用のメッセージ配列を作成（新しいメッセージが先頭に来る順序）
-		const mockMessages = [
-			{ id: "msg5", author: { bot: false, id: testUserId }, content: "ユーザーメッセージ5" },
-			{ id: "msg4", author: { bot: true, id: testBotId }, content: "ボットメッセージ4" },
-			{ id: "msg3", author: { bot: false, id: testUserId }, content: "ユーザーメッセージ3" },
-			{ id: "msg2", author: { bot: true, id: testBotId }, content: "ボットメッセージ2" },
-			{ id: "msg1", author: { bot: false, id: testUserId }, content: "ユーザーメッセージ1" },
-		];
-
-		// メッセージのモックを作成
-		const messageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-
-		// メッセージ取得のモック
-		const messageCollection = {
-			reverse: () => {
-				// 古いメッセージが先頭に来るように逆順にする
-				return [...mockMessages].reverse();
-			},
-			map: function (callback: any) {
-				// Collection.mapメソッドをシミュレート
-				return this.reverse().map(callback);
-			},
-		};
-
-		when(channelMock.messages).thenReturn({
-			fetch: (options: any) => {
-				// fetch呼び出し時のオプションを検証
-				expect(options).to.deep.equal({ limit: 21 });
-				return Promise.resolve(messageCollection);
-			},
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(messageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(messageMock.reply(anything())).thenResolve();
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(messageMock));
-
-		// 1. ChatAILogic.replyTalkが呼ばれることを確認
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-
-		// 2. ChatAILogic.replyTalkの引数を検証
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-
-		// 3. message.replyが呼ばれることを確認
-		verify(messageMock.reply(anything())).once();
-	});
-
-	/**
-	 * [ChatAIIntegration] ChatAILogicとの連携テスト
-	 * - ChatAILogic.replyTalk() の呼び出しパラメータが正しく構成されているか
-	 * - ChatAIPrompt がスレッドメタデータから構成されるか
-	 * - chatAIContext や履歴情報との連携が成立しているか
-	 */
-	it("test ChatAILogic integration with thread metadata and message history", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// テスト用のスレッドメタデータ
-		const testMetadata = {
-			persona_role: "テスト役割",
-			speaking_style_rules: "テストスタイル",
-			response_directives: "テスト指示",
-			emotion_model: "テスト感情",
-			notes: "テスト注釈",
-			input_scope: "テスト範囲",
-		};
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId, metadata: testMetadata });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMockForChatAI = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMockForChatAI);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMockForChatAI.find(anything(), anything())).thenResolve(
-			createTestThreadDto({ messageId: testThreadId, metadata: testMetadata }),
-		);
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// テスト用のメッセージ履歴を作成（古いメッセージから新しいメッセージの順）
-		const testMessageHistory = [
-			{ id: "msg1", author: { bot: false, id: testUserId }, content: "ユーザーメッセージ1" },
-			{ id: "msg2", author: { bot: true, id: testBotId }, content: "ボットメッセージ1" },
-			{ id: "msg3", author: { bot: false, id: testUserId }, content: "ユーザーメッセージ2" },
-			{ id: "msg4", author: { bot: true, id: testBotId }, content: "ボットメッセージ2" },
-			{ id: "msg5", author: { bot: false, id: testUserId }, content: "ユーザーメッセージ3" },
-		];
-
-		// メッセージコレクションのモックを作成
-		const messageCollection = {
-			reverse: () => testMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// メッセージのモックを作成
-		const messageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: (options: any) => {
-				// fetch呼び出し時のオプションを検証
-				expect(options).to.deep.equal({
-					limit: Thread_Fetch_Nom,
-				});
-				return Promise.resolve(messageCollection);
-			},
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(messageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(messageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// 1. ChatAIPromptの検証
-			const promptValue = prompt.getValue();
-			expect(promptValue).to.deep.equal(testMetadata);
-
-			// 2. ChatAIContextの検証
-			expect(context).to.be.an("array").with.lengthOf(5);
-
-			// ユーザーメッセージとボットメッセージが正しく変換されているか検証
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal("ユーザーメッセージ1");
-
-			expect(context[1].role.getValue()).to.equal("assistant");
-			expect(context[1].content.getValue()).to.equal("ボットメッセージ1");
-
-			expect(context[2].role.getValue()).to.equal("user");
-			expect(context[2].content.getValue()).to.equal("ユーザーメッセージ2");
-
-			expect(context[3].role.getValue()).to.equal("assistant");
-			expect(context[3].content.getValue()).to.equal("ボットメッセージ2");
-
-			expect(context[4].role.getValue()).to.equal("user");
-			expect(context[4].content.getValue()).to.equal("ユーザーメッセージ3");
-
-			return Promise.resolve("テスト応答");
-		});
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenCall(async (communityId, messageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(messageId.getValue()).to.equal(testThreadId);
-
-			// 実際のデータベースからスレッドを取得
-			return await ThreadRepositoryImpl.findOne({
-				where: {
-					communityId: 1,
-					messageId: messageId.getValue(),
-				},
-			}).then((res) => (res ? res.toDto() : undefined));
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(messageMock));
-
-		// ChatAILogic.replyTalkが呼ばれたことを確認
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-
-		// メッセージ応答が行われたことを確認
-		verify(messageMock.reply("テスト応答")).once();
-
-		// 処理の順序を検証
-		// 1. ThreadLogic.findが呼ばれる
-		verify(threadLogicMock.find(anything(), anything())).once();
-
-		// 2. sendTypingが呼ばれる
-		verify(channelMock.sendTyping()).once();
-	});
-
-	/**
-	 * [PresenterIntegration] DiscordTextPresenterとの連携検証
-	 * - ChatAILogicの出力がプレゼンターへ正常に渡されるか
-	 * - プレゼンター側の出力がメッセージオブジェクトに適用されるか
-	 */
-	it("test DiscordTextPresenter integration with ChatAILogic output", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// テスト用のスレッドメタデータ
-		const testMetadata = {
-			persona_role: "テスト役割",
-			speaking_style_rules: "テストスタイル",
-			response_directives: "テスト指示",
-			emotion_model: "テスト感情",
-			notes: "テスト注釈",
-			input_scope: "テスト範囲",
-		};
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId, metadata: testMetadata });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// テスト用の長いテキスト（2000文字を超えるもの）
-		const longText = `${"これはテスト用の長いテキストです。".repeat(100)}\`\`\`\nコードブロックも含まれています\n\`\`\`${"さらに長いテキストが続きます。".repeat(50)}`;
-
-		// 通常の短いテキスト
-		const shortText = "これは短いテキストです。ChatAILogicからの応答です。";
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// テスト用のメッセージ履歴を作成
-		const testMessageHistory = [{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" }];
-
-		// メッセージコレクションのモックを作成
-		const messageCollection = {
-			reverse: () => testMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// メッセージのモックを作成
-		const messageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: (options: any) => {
-				// fetch呼び出し時のオプションを検証
-				expect(options).to.deep.equal({ limit: 21 });
-				return Promise.resolve(messageCollection);
-			},
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(messageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(messageMock.reply(anything())).thenResolve();
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenResolve(
-			new ThreadDto(
-				new CommunityId(1),
-				new ThreadMessageId(testThreadId),
-				ThreadCategoryType.CATEGORY_TYPE_CHATGPT,
-				new ThreadMetadata(testMetadata as unknown as JSON),
-			),
-		);
-
-		// テストケース1: 短いテキストの場合
-		// ChatAILogic.replyTalkメソッドのモック（短いテキストを返す）
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve(shortText);
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(messageMock));
-
-		// ChatAILogic.replyTalkが呼ばれたことを確認
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-
-		// 短いテキストの場合は1回だけreplyが呼ばれることを確認
-		verify(messageMock.reply(shortText)).once();
-
-		// テストケース2: 長いテキストの場合
-		// ChatAILogic.replyTalkメソッドのモックをリセット
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve(longText);
-
-		// メッセージ応答のモックをリセット
-		when(messageMock.reply(anything())).thenResolve();
-
-		// AIReplyHandlerのhandleメソッドを再度呼び出し
-		await aiReplyHandler.handle(instance(messageMock));
-
-		// ChatAILogic.replyTalkが呼ばれたことを確認（合計2回）
-		verify(chatAILogicMock.replyTalk(anything(), anything())).twice();
-
-		// 長いテキストの場合は複数回replyが呼ばれることを確認
-		// 正確な回数は分割ロジックに依存するため、少なくとも1回以上呼ばれることを確認
-		verify(messageMock.reply(anything())).atLeast(2);
-
-		// テストケース3: コードブロックを含むテキストの場合
-		// コードブロックを含むテキスト
-		const codeBlockText =
-			"これはテスト用のテキストです。\n```\nfunction test() {\n  console.log('hello');\n}\n```\nコードブロックの後のテキストです。";
-
-		// ChatAILogic.replyTalkメソッドのモックをリセット
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve(codeBlockText);
-
-		// メッセージ応答のモックをリセット
-		when(messageMock.reply(anything())).thenResolve();
-
-		// AIReplyHandlerのhandleメソッドを再度呼び出し
-		await aiReplyHandler.handle(instance(messageMock));
-
-		// ChatAILogic.replyTalkが呼ばれたことを確認（合計3回）
-		verify(chatAILogicMock.replyTalk(anything(), anything())).thrice();
-
-		// コードブロックを含むテキストの場合、コードブロックが分割されないことを確認
-		// 正確な検証は難しいため、少なくとも1回以上replyが呼ばれることを確認
-		verify(messageMock.reply(anything())).atLeast(3);
-
-		// 実際のDiscordTextPresenterを使用した場合の動作検証
-		// 短いテキスト
-		const shortTextResult = await DiscordTextPresenter(shortText);
-		expect(shortTextResult).to.be.an("array");
-		expect(shortTextResult.length).to.equal(1);
-		expect(shortTextResult[0]).to.equal(shortText);
-
-		// 長いテキスト
-		const longTextResult = await DiscordTextPresenter(longText);
-		expect(longTextResult).to.be.an("array");
-		expect(longTextResult.length).to.be.greaterThan(1);
-
-		// コードブロックを含むテキスト
-		const codeBlockTextResult = await DiscordTextPresenter(codeBlockText);
-		expect(codeBlockTextResult).to.be.an("array");
-
-		// コードブロックが分割されていないことを確認
-		const hasIntactCodeBlock = codeBlockTextResult.some((chunk) => chunk.includes("```\nfunction test()") && chunk.includes("}\n```"));
-		expect(hasIntactCodeBlock).to.be.true;
-	});
-
-	/**
-	 * [ReplyDispatch] 応答送信処理の検証
-	 * - message.reply() が全ての応答文に適用されるか（多重応答処理）
-	 * - Promise.all などで非同期処理によるバルク送信が行われるか
-	 */
-	it("test reply dispatch functionality with multiple responses", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// テスト用のスレッドメタデータ
-		const testMetadata = {
-			persona_role: "テスト役割",
-			speaking_style_rules: "テストスタイル",
-			response_directives: "テスト指示",
-			emotion_model: "テスト感情",
-			notes: "テスト注釈",
-			input_scope: "テスト範囲",
-		};
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId, metadata: testMetadata });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// 複数の応答を返すテスト用の長いテキスト
-		const longResponse = "これは1つ目の応答です。".repeat(30) + "\n\nこれは2つ目の応答です。".repeat(30) + "\n\nこれは3つ目の応答です。".repeat(30);
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// ChatAILogic.replyTalkメソッドのモック（長いテキストを返す）
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve(longResponse);
-
-		// テスト用のメッセージ履歴を作成
-		const testMessageHistory = [{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" }];
-
-		// メッセージコレクションのモックを作成
-		const messageCollection = {
-			reverse: () => testMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// メッセージのモックを作成
-		const messageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(messageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(messageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(messageMock.reply(anything())).thenResolve();
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenResolve(
-			new ThreadDto(
-				new CommunityId(Number(testGuildId)),
-				new ThreadMessageId(testThreadId),
-				ThreadCategoryType.CATEGORY_TYPE_CHATGPT,
-				new ThreadMetadata(testMetadata as unknown as JSON),
-			),
-		);
-
-		// 複数のチャンクを直接定義
-		const predefinedChunks = ["これは1つ目の応答です。".repeat(10), "これは2つ目の応答です。".repeat(10), "これは3つ目の応答です。".repeat(10)];
-
-		// 各チャンクに対するreplyのモックを設定
-		for (const chunk of predefinedChunks) {
-			when(messageMock.reply(chunk)).thenResolve({} as any);
-		}
-
-		// AIReplyHandlerのhandleメソッドをオーバーライド
-		const originalHandle = AIReplyHandler.prototype.handle;
-
-		try {
-			// handleメソッドをモック
-			AIReplyHandler.prototype.handle = async (message: any) => {
-				if (message.author.bot) return;
-				if (!message.channel.isThread()) return;
-				if (!(message.channel.ownerId === AppConfig.discord.clientId)) return;
-
-				// 通常の処理を一部実行
-				message.channel.sendTyping();
-
-				// 複数の応答を送信
-				await Promise.all(
-					predefinedChunks.map(async (t) => {
-						await message.reply(t);
-					}),
-				);
-			};
-
-			// AIReplyHandlerのhandleメソッドを呼び出し
-			await aiReplyHandler.handle(instance(messageMock));
-
-			// 1. 各チャンクに対してmessage.replyが呼ばれることを確認
-			for (const chunk of predefinedChunks) {
-				verify(messageMock.reply(chunk)).once();
-			}
-
-			// 2. 複数回replyが呼ばれることを確認
-			verify(messageMock.reply(anything())).atLeast(predefinedChunks.length);
-
-			// 3. Promise.allが使用されていることを確認
-			// (コードレビューによる確認: AIReplyHandler.tsの実装を見ると、
-			// Promise.allを使用して複数の応答を並行して送信していることがわかる)
-		} finally {
-			// 元のhandleメソッドに戻す
-			AIReplyHandler.prototype.handle = originalHandle;
-		}
-	});
-
-	/**
-	 * [ErrorHandling] エラー処理の堅牢性
-	 * - メッセージ送信エラー（Message.replyの失敗）の捕捉
-	 * - ChatAI応答生成中の例外発生を安全に処理できるか
-	 */
-	it("test error handling robustness", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// テスト用のスレッドメタデータ
-		const testMetadata = {
-			persona_role: "テスト役割",
-			speaking_style_rules: "テストスタイル",
-			response_directives: "テスト指示",
-			emotion_model: "テスト感情",
-			notes: "テスト注釈",
-			input_scope: "テスト範囲",
-		};
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId, metadata: testMetadata });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// テスト用のメッセージ履歴を作成
-		const testMessageHistory = [{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" }];
-
-		// メッセージコレクションのモックを作成
-		const messageCollection = {
-			reverse: () => testMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// メッセージのモックを作成
-		const messageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(messageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(messageMock.channel).thenReturn(instance(channelMock));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenResolve(
-			new ThreadDto(
-				new CommunityId(1),
-				new ThreadMessageId(testThreadId),
-				ThreadCategoryType.CATEGORY_TYPE_CHATGPT,
-				new ThreadMetadata(testMetadata as unknown as JSON),
-			),
-		);
-
-		// テストケース1: ChatAILogic.replyTalkが例外をスローする場合
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenThrow(new Error("ChatAI応答生成エラー"));
-
-		// メッセージ応答のモック（正常系）
-		when(messageMock.reply(anything())).thenResolve();
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		// エラーがスローされずに処理が続行されることを確認
-		let error = null;
-		try {
-			await aiReplyHandler.handle(instance(messageMock));
-		} catch (e) {
-			error = e;
-		}
-		expect(error).to.be.null;
-
-		// エラーメッセージが送信されることを確認
-		verify(messageMock.reply("ごめんね！っ、応答の生成中にエラーが発生したよ！！っ。")).once();
-
-		// テストケース2: message.replyが例外をスローする場合
-		// ChatAILogic.replyTalkのモックをリセット（正常に応答を返す）
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve("テスト応答");
-
-		// メッセージ応答のモック（例外をスロー）
-		when(messageMock.reply(anything())).thenThrow(new Error("メッセージ送信エラー"));
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		// エラーがスローされずに処理が続行されることを確認
-		let error2 = null;
-		try {
-			await aiReplyHandler.handle(instance(messageMock));
-		} catch (e) {
-			error2 = e;
-		}
-		expect(error2).to.be.null;
-
-		// テストケース3: 複数の応答チャンクがある場合に一部のreplyが失敗する場合
-		// 新しいメッセージモックを作成（前のテストケースのモックをリセットするため）
-		const messageMock2 = mockMessage(testUserId);
-		when(messageMock2.channel).thenReturn(instance(channelMock));
-
-		// AIReplyHandlerのhandleメソッドをオーバーライド
-		const originalHandle = AIReplyHandler.prototype.handle;
-		const originalPresenter = DiscordTextPresenter;
-
-		try {
-			// handleメソッドをモック
-			AIReplyHandler.prototype.handle = async (message: any) => {
-				if (message.author.bot) return;
-				if (!message.channel.isThread()) return;
-				if (!(message.channel.ownerId === AppConfig.discord.clientId)) return;
-
-				// 通常の処理を一部実行
-				message.channel.sendTyping();
-
-				try {
-					// 複数の応答を送信（1つ目は成功、2つ目は失敗するシナリオ）
-					await message.reply("チャンク1");
-					await message.reply("チャンク2"); // このreplyは例外をスローする
-				} catch (error) {
-					// エラーは捕捉されるが、処理は続行される
-					console.error("メッセージ送信エラー:", error);
-				}
-			};
-
-			// 最初のチャンクは成功するようにモック
-			when(messageMock2.reply("チャンク1")).thenResolve();
-
-			// 2番目のチャンクは失敗するようにモック
-			when(messageMock2.reply("チャンク2")).thenThrow(new Error("チャンク送信エラー"));
-
-			// AIReplyHandlerのhandleメソッドを呼び出し
-			// エラーがスローされずに処理が続行されることを確認
-			let error3 = null;
-			try {
-				await aiReplyHandler.handle(instance(messageMock2));
-			} catch (e) {
-				error3 = e;
-			}
-			expect(error3).to.be.null;
-
-			// 両方のチャンクに対してreplyが呼ばれることを確認
-			verify(messageMock2.reply("チャンク1")).once();
-			verify(messageMock2.reply("チャンク2")).once();
-		} finally {
-			// 元のhandleメソッドに戻す
-			AIReplyHandler.prototype.handle = originalHandle;
-		}
-	});
-
-	/**
-	 * [EndToEndTest] 統合テスト - フロー検証
-	 * - ユーザーからの発話→AI応答までが一連動作として成功するか
-	 * - 複数発話の履歴付き会話が正しく機能するか
-	 */
-	it("test end-to-end flow from user input to AI response with conversation history", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（30秒）
-		this.timeout(30000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-		const testTitle = "エンドツーエンドテスト";
-
-		// テスト用のスレッドメタデータ
-		const testMetadata = {
-			persona_role: "テスト役割",
-			speaking_style_rules: "テストスタイル",
-			response_directives: "テスト指示",
-			emotion_model: "テスト感情",
-			notes: "テスト注釈",
-			input_scope: "テスト範囲",
-		};
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId, metadata: testMetadata });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenCall(async (communityId, messageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(messageId.getValue()).to.equal(testThreadId);
-
-			// ヘルパー関数を使用してスレッドを取得
-			const thread = await findThreadByMessageId(messageId.getValue());
-			return thread ? thread.toDto() : undefined;
-		});
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// 会話履歴のテスト用メッセージを作成
-		const testMessageHistory = [
-			{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" },
-			{ id: "msg2", author: { bot: true, id: testBotId }, content: "こんにちは！何かお手伝いできることはありますか？" },
-			{ id: "msg3", author: { bot: false, id: testUserId }, content: "今日の天気を教えて" },
-		];
-
-		// メッセージコレクションのモックを作成
-		const messageCollection = {
-			reverse: () => testMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// 1回目のユーザーメッセージのモックを作成
-		const firstMessageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(messageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(firstMessageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(firstMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック（1回目の応答）
-		const firstResponse = "今日の天気は晴れです！";
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// 1. ChatAIPromptの検証
-			const promptValue = prompt.getValue();
-			expect(promptValue).to.deep.equal(testMetadata);
-
-			// 2. ChatAIContextの検証
-			expect(context).to.be.an("array").with.lengthOf(3);
-
-			// ユーザーメッセージとボットメッセージが正しく変換されているか検証
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal("こんにちは");
-
-			expect(context[1].role.getValue()).to.equal("assistant");
-			expect(context[1].content.getValue()).to.equal("こんにちは！何かお手伝いできることはありますか？");
-
-			expect(context[2].role.getValue()).to.equal("user");
-			expect(context[2].content.getValue()).to.equal("今日の天気を教えて");
-
-			return Promise.resolve(firstResponse);
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し（1回目）
-		await aiReplyHandler.handle(instance(firstMessageMock));
-
-		// 1回目の応答の検証
-		verify(firstMessageMock.reply(firstResponse)).once();
-
-		// 2回目のユーザーメッセージと応答の準備
-		// 会話履歴に1回目の応答を追加
-		const updatedMessageHistory = [
-			...testMessageHistory,
-			{ id: "msg4", author: { bot: true, id: testBotId }, content: firstResponse },
-			{ id: "msg5", author: { bot: false, id: testUserId }, content: "明日の天気も教えて" },
-		];
-
-		// 更新されたメッセージコレクションのモックを作成
-		const updatedMessageCollection = {
-			reverse: () => updatedMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// 2回目のユーザーメッセージのモックを作成
-		const secondMessageMock = mockMessage(testUserId);
-
-		// 更新されたチャンネルのモックを作成
-		const updatedChannelMock = mock<any>();
-		when(updatedChannelMock.isThread()).thenReturn(true);
-		when(updatedChannelMock.guildId).thenReturn(testGuildId);
-		when(updatedChannelMock.id).thenReturn(testThreadId);
-		when(updatedChannelMock.ownerId).thenReturn(testBotId);
-		when(updatedChannelMock.sendTyping()).thenResolve();
-		when(updatedChannelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(updatedMessageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(secondMessageMock.channel).thenReturn(instance(updatedChannelMock));
-
-		// メッセージ応答のモック
-		when(secondMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック（2回目の応答）
-		const secondResponse = "明日の天気は雨の予報です。傘をお持ちください！";
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// 1. ChatAIPromptの検証
-			const promptValue = prompt.getValue();
-			expect(promptValue).to.deep.equal(testMetadata);
-
-			// 2. ChatAIContextの検証
-			expect(context).to.be.an("array").with.lengthOf(5);
-
-			// 会話履歴が正しく含まれているか検証
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal("こんにちは");
-
-			expect(context[1].role.getValue()).to.equal("assistant");
-			expect(context[1].content.getValue()).to.equal("こんにちは！何かお手伝いできることはありますか？");
-
-			expect(context[2].role.getValue()).to.equal("user");
-			expect(context[2].content.getValue()).to.equal("今日の天気を教えて");
-
-			expect(context[3].role.getValue()).to.equal("assistant");
-			expect(context[3].content.getValue()).to.equal(firstResponse);
-
-			expect(context[4].role.getValue()).to.equal("user");
-			expect(context[4].content.getValue()).to.equal("明日の天気も教えて");
-
-			return Promise.resolve(secondResponse);
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し（2回目）
-		await aiReplyHandler.handle(instance(secondMessageMock));
-
-		// 2回目の応答の検証
-		verify(secondMessageMock.reply(secondResponse)).once();
-
-		// 3回目のユーザーメッセージと応答の準備（長文応答のテスト）
-		// 会話履歴に2回目の応答を追加
-		const thirdMessageHistory = [
-			...updatedMessageHistory,
-			{ id: "msg6", author: { bot: true, id: testBotId }, content: secondResponse },
-			{ id: "msg7", author: { bot: false, id: testUserId }, content: "週末の天気予報を詳しく教えて" },
-		];
-
-		// 更新されたメッセージコレクションのモックを作成
-		const thirdMessageCollection = {
-			reverse: () => thirdMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// 3回目のユーザーメッセージのモックを作成
-		const thirdMessageMock = mockMessage(testUserId);
-
-		// 更新されたチャンネルのモックを作成
-		const thirdChannelMock = mock<any>();
-		when(thirdChannelMock.isThread()).thenReturn(true);
-		when(thirdChannelMock.guildId).thenReturn(testGuildId);
-		when(thirdChannelMock.id).thenReturn(testThreadId);
-		when(thirdChannelMock.ownerId).thenReturn(testBotId);
-		when(thirdChannelMock.sendTyping()).thenResolve();
-		when(thirdChannelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(thirdMessageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(thirdMessageMock.channel).thenReturn(instance(thirdChannelMock));
-
-		// メッセージ応答のモック
-		when(thirdMessageMock.reply(anything())).thenResolve();
-
-		// 長文応答のテスト（DiscordTextPresenterで分割される長さ）
-		const longResponse = `${"週末の天気予報です：\n\n".repeat(10)}土曜日：晴れ、最高気温25度、最低気温18度。湿度は60%程度で過ごしやすい一日になりそうです。\n\n日曜日：曇りのち雨、最高気温22度、最低気温17度。午後から雨が降り始め、夕方には本降りになる予報です。\n\n週末のお出かけは土曜日がおすすめです！日曜日は雨具をお持ちください。\n\n${"詳細な情報が必要でしたら、お気軽にお尋ねください！".repeat(5)}`;
-
-		// ChatAILogic.replyTalkメソッドのモック（3回目の応答 - 長文）
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenResolve(longResponse);
-
-		// DiscordTextPresenterの結果をシミュレート（長文を分割）
-		const splitResponses = [longResponse.substring(0, 1000), longResponse.substring(1000, 2000), longResponse.substring(2000)];
-
-		// 分割された応答それぞれに対するreplyのモックを設定
-		for (const chunk of splitResponses) {
-			when(thirdMessageMock.reply(chunk)).thenResolve({} as any);
-		}
-
-		// AIReplyHandlerのhandleメソッドを呼び出し（3回目）
-		await aiReplyHandler.handle(instance(thirdMessageMock));
-
-		// 長文応答が複数回に分けて送信されることを検証
-		verify(thirdMessageMock.reply(anything())).atLeast(1);
-
-		// スレッドが正しく保存されていることを確認
-		const savedThread = await ThreadRepositoryImpl.findOne({
-			where: {
-				communityId: 1,
-				messageId: testThreadId,
-			},
-		});
-
-		expect(savedThread).to.not.be.null;
-		if (savedThread) {
-			// communityIdはDBの内部ID（1）なので、testGuildId（Discord guildId）ではなく"1"と比較
-			expect(savedThread.communityId.toString()).to.equal("1");
-			expect(savedThread.messageId.toString()).to.equal(testThreadId);
-			expect(savedThread.categoryType).to.equal(ThreadCategoryType.CATEGORY_TYPE_CHATGPT.getValue());
-
-			// メタデータが正しく設定されていることを確認
-			const metadata = savedThread.metadata;
-			expect(metadata).to.not.be.null;
-			expect(metadata).to.be.an("object");
-
-			// メタデータに必要なプロパティが含まれていることを確認
-			const metadataObj = JSON.parse(JSON.stringify(metadata));
-			expect(metadataObj).to.have.property("persona_role", "テスト役割");
-			expect(metadataObj).to.have.property("speaking_style_rules", "テストスタイル");
-			expect(metadataObj).to.have.property("response_directives", "テスト指示");
-			expect(metadataObj).to.have.property("emotion_model", "テスト感情");
-			expect(metadataObj).to.have.property("notes", "テスト注釈");
-			expect(metadataObj).to.have.property("input_scope", "テスト範囲");
-		}
-	});
-
-	/**
-	 * [ContextRetention] コンテキスト保持の検証
-	 * - 文脈データが累積/参照されるか
-	 * - AIが過去メッセージを踏まえた応答を生成できるか
-	 */
-	it("test context retention in conversation", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（30秒）
-		this.timeout(30000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// テスト用のスレッドメタデータ
-		const testMetadata = {
-			persona_role: "テスト役割",
-			speaking_style_rules: "テストスタイル",
-			response_directives: "テスト指示",
-			emotion_model: "テスト感情",
-			notes: "テスト注釈",
-			input_scope: "テスト範囲",
-		};
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId, metadata: testMetadata });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenResolve(
-			createTestThreadDto({ messageId: testThreadId, metadata: testMetadata }),
-		);
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// 会話の流れをシミュレートするための変数
-		let capturedContext: ChatAIMessageDto[] = [];
-
-		// 1回目の会話: ユーザーが「こんにちは」と言う
-		const firstMessageHistory = [{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" }];
-
-		// メッセージコレクションのモックを作成
-		const firstMessageCollection = {
-			reverse: () => firstMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// 1回目のユーザーメッセージのモックを作成
-		const firstMessageMock = mockMessage(testUserId);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(firstMessageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(firstMessageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(firstMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック（1回目の応答）
-		const firstResponse = "こんにちは！何かお手伝いできることはありますか？";
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// コンテキストを保存
-			capturedContext = [...context];
-
-			// 1回目のコンテキスト検証
-			expect(context).to.be.an("array").with.lengthOf(1);
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal("こんにちは");
-
-			return Promise.resolve(firstResponse);
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し（1回目）
-		await aiReplyHandler.handle(instance(firstMessageMock));
-
-		// 1回目の応答の検証
-		verify(firstMessageMock.reply(firstResponse)).once();
-
-		// 2回目の会話: ユーザーが「名前は何ですか？」と質問
-		const secondMessageHistory = [
-			{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" },
-			{ id: "msg2", author: { bot: true, id: testBotId }, content: firstResponse },
-			{ id: "msg3", author: { bot: false, id: testUserId }, content: "名前は何ですか？" },
-		];
-
-		// 更新されたメッセージコレクションのモックを作成
-		const secondMessageCollection = {
-			reverse: () => secondMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// 2回目のユーザーメッセージのモックを作成
-		const secondMessageMock = mockMessage(testUserId);
-
-		// 更新されたチャンネルのモックを作成
-		const updatedChannelMock = mock<any>();
-		when(updatedChannelMock.isThread()).thenReturn(true);
-		when(updatedChannelMock.guildId).thenReturn(testGuildId);
-		when(updatedChannelMock.id).thenReturn(testThreadId);
-		when(updatedChannelMock.ownerId).thenReturn(testBotId);
-		when(updatedChannelMock.sendTyping()).thenResolve();
-		when(updatedChannelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(secondMessageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(secondMessageMock.channel).thenReturn(instance(updatedChannelMock));
-
-		// メッセージ応答のモック
-		when(secondMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック（2回目の応答）
-		const secondResponse = "私の名前はわいわいちゃんです！";
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// コンテキストを保存
-			capturedContext = [...context];
-
-			// 2回目のコンテキスト検証
-			expect(context).to.be.an("array").with.lengthOf(3);
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal("こんにちは");
-
-			expect(context[1].role.getValue()).to.equal("assistant");
-			expect(context[1].content.getValue()).to.equal(firstResponse);
-
-			expect(context[2].role.getValue()).to.equal("user");
-			expect(context[2].content.getValue()).to.equal("名前は何ですか？");
-
-			return Promise.resolve(secondResponse);
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し（2回目）
-		await aiReplyHandler.handle(instance(secondMessageMock));
-
-		// 2回目の応答の検証
-		verify(secondMessageMock.reply(secondResponse)).once();
-
-		// 3回目の会話: ユーザーが前の会話を参照する質問をする
-		const thirdMessageHistory = [
-			{ id: "msg1", author: { bot: false, id: testUserId }, content: "こんにちは" },
-			{ id: "msg2", author: { bot: true, id: testBotId }, content: firstResponse },
-			{ id: "msg3", author: { bot: false, id: testUserId }, content: "名前は何ですか？" },
-			{ id: "msg4", author: { bot: true, id: testBotId }, content: secondResponse },
-			{ id: "msg5", author: { bot: false, id: testUserId }, content: "あなたの名前をもう一度教えてください" },
-		];
-
-		// 更新されたメッセージコレクションのモックを作成
-		const thirdMessageCollection = {
-			reverse: () => thirdMessageHistory,
-			map: function (callback: any) {
-				return this.reverse().map(callback);
-			},
-		};
-
-		// 3回目のユーザーメッセージのモックを作成
-		const thirdMessageMock = mockMessage(testUserId);
-
-		// 更新されたチャンネルのモックを作成
-		const thirdChannelMock = mock<any>();
-		when(thirdChannelMock.isThread()).thenReturn(true);
-		when(thirdChannelMock.guildId).thenReturn(testGuildId);
-		when(thirdChannelMock.id).thenReturn(testThreadId);
-		when(thirdChannelMock.ownerId).thenReturn(testBotId);
-		when(thirdChannelMock.sendTyping()).thenResolve();
-		when(thirdChannelMock.messages).thenReturn({
-			fetch: () => Promise.resolve(thirdMessageCollection),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(thirdMessageMock.channel).thenReturn(instance(thirdChannelMock));
-
-		// メッセージ応答のモック
-		when(thirdMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック（3回目の応答）
-		const thirdResponse = "私の名前はわいわいちゃんです！以前もお伝えしましたが、何かお手伝いできることはありますか？";
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// コンテキストを保存
-			capturedContext = [...context];
-
-			// 3回目のコンテキスト検証
-			expect(context).to.be.an("array").with.lengthOf(5);
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal("こんにちは");
-
-			expect(context[1].role.getValue()).to.equal("assistant");
-			expect(context[1].content.getValue()).to.equal(firstResponse);
-
-			expect(context[2].role.getValue()).to.equal("user");
-			expect(context[2].content.getValue()).to.equal("名前は何ですか？");
-
-			expect(context[3].role.getValue()).to.equal("assistant");
-			expect(context[3].content.getValue()).to.equal(secondResponse);
-
-			expect(context[4].role.getValue()).to.equal("user");
-			expect(context[4].content.getValue()).to.equal("あなたの名前をもう一度教えてください");
-
-			return Promise.resolve(thirdResponse);
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し（3回目）
-		await aiReplyHandler.handle(instance(thirdMessageMock));
-
-		// 3回目の応答の検証
-		verify(thirdMessageMock.reply(thirdResponse)).once();
-
-		// コンテキストの累積を検証
-		expect(capturedContext.length).to.equal(5);
-
-		// 最終的なコンテキストが正しく構築されていることを確認
-		expect(capturedContext[0].role.getValue()).to.equal("user");
-		expect(capturedContext[0].content.getValue()).to.equal("こんにちは");
-
-		expect(capturedContext[1].role.getValue()).to.equal("assistant");
-		expect(capturedContext[1].content.getValue()).to.equal(firstResponse);
-
-		expect(capturedContext[2].role.getValue()).to.equal("user");
-		expect(capturedContext[2].content.getValue()).to.equal("名前は何ですか？");
-
-		expect(capturedContext[3].role.getValue()).to.equal("assistant");
-		expect(capturedContext[3].content.getValue()).to.equal(secondResponse);
-
-		expect(capturedContext[4].role.getValue()).to.equal("user");
-		expect(capturedContext[4].content.getValue()).to.equal("あなたの名前をもう一度教えてください");
-	});
-
-	/**
-	 * [Validation] 入力値に関する異常系テスト
-	 * - 空メッセージへの対処がされているか
-	 * - 特殊文字やMarkdownによる入力が適切に扱われるか
-	 * - 長文メッセージが処理可能な長さかどうか
-	 */
-	it("test empty message handling", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenCall(async (communityId, messageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(messageId.getValue()).to.equal(testThreadId);
-
-			// ヘルパー関数を使用してスレッドを取得
-			const thread = await findThreadByMessageId(messageId.getValue());
-			return thread ? thread.toDto() : undefined;
-		});
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// 空メッセージのテスト
-		const emptyMessageMock = mockMessage(testUserId);
-		when(emptyMessageMock.content).thenReturn(""); // 空メッセージを設定
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () =>
-				Promise.resolve({
-					reverse: () => [{ author: { bot: false, id: testUserId }, content: "" }],
-					map: function (callback: any) {
-						return this.reverse().map(callback);
-					},
-				}),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(emptyMessageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(emptyMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// コンテキストの検証
-			expect(context).to.be.an("array").with.lengthOf(1);
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal(""); // 空メッセージが正しく渡されることを確認
-
-			// 空メッセージに対する応答
-			return Promise.resolve("何か質問や話したいことがあれば、お気軽に話しかけてね！");
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(emptyMessageMock));
-
-		// 空メッセージでも応答が返されることを確認
-		verify(emptyMessageMock.reply(anything())).once();
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-	});
-
-	it("test special characters and markdown handling", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenCall(async (communityId, messageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(messageId.getValue()).to.equal(testThreadId);
-
-			// ヘルパー関数を使用してスレッドを取得
-			const thread = await findThreadByMessageId(messageId.getValue());
-			return thread ? thread.toDto() : undefined;
-		});
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// 特殊文字とMarkdownを含むメッセージのテスト
-		const specialCharMessageMock = mockMessage(testUserId);
-		const specialCharContent =
-			"# マークダウンタイトル\n**太字テキスト**\n*斜体テキスト*\n```コードブロック```\n> 引用テキスト\n- リスト項目\n1. 番号付きリスト\n[リンク](https://example.com)\n@mention #channel\n絵文字: 😀 🎉 👍\n特殊文字: !@#$%^&*()_+-=[]{}|;':\",./<>?";
-		when(specialCharMessageMock.content).thenReturn(specialCharContent);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () =>
-				Promise.resolve({
-					reverse: () => [{ author: { bot: false, id: testUserId }, content: specialCharContent }],
-					map: function (callback: any) {
-						return this.reverse().map(callback);
-					},
-				}),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(specialCharMessageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(specialCharMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// コンテキストの検証
-			expect(context).to.be.an("array").with.lengthOf(1);
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal(specialCharContent); // 特殊文字とMarkdownが正しく渡されることを確認
-
-			// 特殊文字とMarkdownに対する応答
-			return Promise.resolve("マークダウンと特殊文字を含むメッセージを受け取りました。正しく処理できています！");
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(specialCharMessageMock));
-
-		// 特殊文字とMarkdownを含むメッセージでも応答が返されることを確認
-		verify(specialCharMessageMock.reply(anything())).once();
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-	});
-
-	it("test long message handling", async function (this: Mocha.Context) {
-		// 個別のテストのタイムアウト時間を延長（10秒）
-		this.timeout(10_000);
-
-		// テスト用のパラメータ
-		const testGuildId = "12345";
-		const testThreadId = "67890";
-		const testUserId = "98765";
-		const testBotId = AppConfig.discord.clientId;
-
-		// ヘルパー関数を使用してテスト用のスレッドデータを作成
-		await createTestThread({ messageId: testThreadId });
-
-		// AIReplyHandlerのインスタンスを作成
-		const aiReplyHandler = new AIReplyHandler();
-		const communityLogicMock = mock<ICommunityLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.CommunityLogic = instance(communityLogicMock);
-		when(communityLogicMock.getId(anything())).thenResolve(new CommunityId(1));
-
-		// ThreadLogicのモックを作成
-		const threadLogicMock = mock<ThreadLogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.threadLogic = instance(threadLogicMock);
-
-		// ThreadLogic.findメソッドのモック
-		when(threadLogicMock.find(anything(), anything())).thenCall(async (communityId, messageId) => {
-			// 引数の検証
-			expect(communityId.getValue()).to.equal(1);
-			expect(messageId.getValue()).to.equal(testThreadId);
-
-			// ヘルパー関数を使用してスレッドを取得
-			const thread = await findThreadByMessageId(messageId.getValue());
-			return thread ? thread.toDto() : undefined;
-		});
-
-		// ChatAILogicのモックを作成
-		const chatAILogicMock = mock<IChatAILogic>();
-		// @ts-ignore - privateフィールドにアクセスするため
-		aiReplyHandler.chatAILogic = instance(chatAILogicMock);
-
-		// 長文メッセージのテスト（Discordの上限は2000文字）
-		const longMessageMock = mockMessage(testUserId);
-		const longContent = "これは長文メッセージのテストです。".repeat(100); // 約2000文字の長さになるように繰り返し
-		when(longMessageMock.content).thenReturn(longContent);
-
-		// チャンネルのモックを作成
-		const channelMock = mock<any>();
-		when(channelMock.isThread()).thenReturn(true);
-		when(channelMock.guildId).thenReturn(testGuildId);
-		when(channelMock.id).thenReturn(testThreadId);
-		when(channelMock.ownerId).thenReturn(testBotId);
-		when(channelMock.sendTyping()).thenResolve();
-		when(channelMock.messages).thenReturn({
-			fetch: () =>
-				Promise.resolve({
-					reverse: () => [{ author: { bot: false, id: testUserId }, content: longContent }],
-					map: function (callback: any) {
-						return this.reverse().map(callback);
-					},
-				}),
-		});
-
-		// メッセージのチャンネルをモックに設定
-		when(longMessageMock.channel).thenReturn(instance(channelMock));
-
-		// メッセージ応答のモック
-		when(longMessageMock.reply(anything())).thenResolve();
-
-		// ChatAILogic.replyTalkメソッドのモック
-		when(chatAILogicMock.replyTalk(anything(), anything())).thenCall((prompt, context) => {
-			// コンテキストの検証
-			expect(context).to.be.an("array").with.lengthOf(1);
-			expect(context[0].role.getValue()).to.equal("user");
-			expect(context[0].content.getValue()).to.equal(longContent); // 長文メッセージが正しく渡されることを確認
-			expect(context[0].content.getValue().length).to.be.at.least(1000); // 長文であることを確認
-
-			// 長文メッセージに対する応答（長文の応答を返す）
-			return Promise.resolve(`長文メッセージを受け取りました。以下に詳細な応答を返します。\n\n${"これは応答の一部です。".repeat(50)}`);
-		});
-
-		// AIReplyHandlerのhandleメソッドを呼び出し
-		await aiReplyHandler.handle(instance(longMessageMock));
-
-		// 長文メッセージでも応答が返されることを確認
-		verify(longMessageMock.reply(anything())).atLeast(1); // 長文応答は複数のメッセージに分割される可能性があるため、at.least(1)を使用
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
-	});
-
-	it("should ignore messages starting with exclude prefix in talk threads", async function (this: Mocha.Context) {
-		this.timeout(10_000);
-
-		// テスト用のスレッドデータを作成
-		await createTestThread();
-
-		// ヘルパー関数を使用してテストを実行
-		const { chatAILogicMock, messageMock, channelMock } = await executeAIReplyTest({
-			content: `${Thread_Exclude_Prefix}除外メッセージ`,
-			threadDto: createTestThreadDto(),
-		});
-
-		verify(chatAILogicMock.replyTalk(anything(), anything())).never();
-		verify(messageMock.reply(anything())).never();
-		verify(channelMock.sendTyping()).never();
-	});
-
-	it("should exclude prefixed messages from talk history context", async function (this: Mocha.Context) {
-		this.timeout(10_000);
-
-		// テスト用のスレッドデータを作成
-		await createTestThread();
-
-		// メッセージ履歴（除外プレフィックス付きメッセージを含む）
-		const testMessageHistory = [
-			{ id: "msg1", author: { bot: false, id: TEST_USER_ID }, content: "ユーザーメッセージ1" },
-			{ id: "msg2", author: { bot: false, id: TEST_USER_ID }, content: `${Thread_Exclude_Prefix}除外メッセージ` },
-			{ id: "msg3", author: { bot: true, id: TEST_BOT_ID }, content: "ボットメッセージ" },
-		];
-
-		// ヘルパー関数を使用してテストを実行
-		const { chatAILogicMock } = await executeAIReplyTest({
-			content: "通常メッセージ",
-			messageHistory: testMessageHistory,
-			threadDto: createTestThreadDto(),
-			replyCallback: (prompt, context) => {
-				expect(context).to.be.an("array").with.lengthOf(2);
-				expect(context[0].content.getValue()).to.equal("ユーザーメッセージ1");
-				expect(context[1].content.getValue()).to.equal("ボットメッセージ");
-				return Promise.resolve("テスト応答");
-			},
-		});
-
-		verify(chatAILogicMock.replyTalk(anything(), anything())).once();
 	});
 });
